# specs/001-volunteer-care-report/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StrayHub CRM Care Report API
  version: 0.1.0
  description: >-
    志工日常照護回報與動物近期歷程的 CRM 邊界契約。一般收容所操作由
    FastAPI 重新驗證 Session、Membership、角色與 Active Shelter Context；
    PLATFORM_ADMIN 使用平台級 PLATFORM Scope，且所有跨收容所操作都留下 Audit Record。
...
servers:
  - url: http://localhost:8000
    description: 本機開發
...
security:
  - bearerAuth: []
paths:
...
  /v1/organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      responses:
        "200":
          description: 依角色可見的收容所
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationListResponse" }
    post:
      tags: [Organizations]
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrganizationCreateRequest" }
      responses:
        "201":
          description: 收容所已建立，預設未啟用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: getOrganization
      responses:
        "200":
          description: 收容所資料
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "404": { $ref: "#/components/responses/NotFoundOrForbidden" }
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrganizationUpdateRequest" }
      responses:
        "200":
          description: 收容所已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/disable:
    post:
      tags: [Organizations]
      operationId: disableOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "204": { description: 收容所已停用 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/initial-admin:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    post:
      tags: [Organizations]
      operationId: createInitialShelterAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InitialAdminCreateRequest" }
      responses:
        "201":
          description: 初始收容所管理員帳號已建立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/memberships:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Memberships]
      operationId: listMemberships
      responses:
        "200":
          description: 收容所 Membership
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipListResponse" }
    post:
      tags: [Memberships]
      operationId: createMembership
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipCreateRequest" }
      responses:
        "201":
          description: Membership 已建立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/accounts:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    post:
      tags: [Memberships]
      operationId: createOrganizationAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountCreateRequest" }
      responses:
        "201":
          description: 收容所帳號與 Membership 已建立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/memberships/{membershipId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/MembershipId"
    patch:
      tags: [Memberships]
      operationId: updateMembership
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipUpdateRequest" }
      responses:
        "200":
          description: Membership 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/areas:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: listShelterAreas
      responses:
        "200":
          description: 收容所 Cage/Area 清單
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShelterAreaListResponse" }
    post:
      tags: [Organizations]
      operationId: createShelterArea
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShelterAreaCreateRequest" }
      responses:
        "201":
          description: Cage/Area 已建立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShelterArea" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/organizations/{organizationId}/areas/{areaId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - name: areaId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Organizations]
      operationId: updateShelterArea
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShelterAreaUpdateRequest" }
      responses:
        "200":
          description: Cage/Area 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShelterArea" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 簽署的短效 Access Token；每次受保護 Request 仍必須通過 Server-side Session 與 Organization Scope 驗證。
  parameters:
    OrganizationId:
      {
        name: organizationId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
    MembershipId:
      {
        name: membershipId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
  responses:
...
    Forbidden:
      description: 無權限或無法存取，不能洩漏資源存在性
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    NotFoundOrForbidden:
      description: 不存在或未授權的一致結果
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        operation_id: { type: [string, "null"], format: uuid }
        details: { type: object, additionalProperties: true }
...
    Organization:
      type: object
      required: [id, code, name, status, timezone, timezone_version]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [pending_setup, active, suspended] }
        timezone: { type: string, example: Asia/Taipei }
        timezone_version: { type: integer, minimum: 1 }
    OrganizationCreateRequest:
      type: object
      required:
        [
          code,
          name,
          status,
          initial_admin_username,
          initial_admin_temporary_password,
        ]
      properties:
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [pending_setup] }
        initial_admin_username: { type: string, minLength: 1 }
        initial_admin_temporary_password:
          { type: string, format: password, minLength: 1 }
        address: { type: string }
        service_area: { type: string }
        contact: { type: string }
    InitialAdminCreateRequest:
      type: object
      required: [username, temporary_password]
      properties:
        username: { type: string }
        temporary_password: { type: string, format: password }
    OrganizationUpdateRequest:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        service_area: { type: string }
        contact: { type: string }
        status: { type: string, enum: [pending_setup, active, suspended] }
        timezone: { type: string }
    OrganizationListResponse:
      type: object
      required: [items]
      properties:
        items:
          { type: array, items: { $ref: "#/components/schemas/Organization" } }
    Membership:
      type: object
      required: [id, organization_id, user_id, role, status, medical_care_access]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        role: { type: string, enum: [SHELTER_ADMIN, STAFF, VOLUNTEER] }
        status: { type: string, enum: [invited, active, disabled, expired, revoked] }
        valid_from:
          type: [string, "null"]
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
        access_version: { type: integer, minimum: 1, default: 1 }
        medical_care_access: { type: boolean, default: false }
    MembershipCreateRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id: { type: string, format: uuid }
        role: { type: string, enum: [SHELTER_ADMIN, STAFF, VOLUNTEER] }
    MembershipUpdateRequest:
      type: object
      properties:
        role: { type: string, enum: [SHELTER_ADMIN, STAFF, VOLUNTEER] }
        status: { type: string, enum: [invited, active, disabled] }
        medical_care_access: { type: boolean }
    MembershipListResponse:
      type: object
      required: [items]
      properties:
        items:
          { type: array, items: { $ref: "#/components/schemas/Membership" } }
    AccountCreateRequest:
      type: object
      required: [username, display_name, temporary_password, role]
      properties:
        username: { type: string, minLength: 1 }
        display_name: { type: string, minLength: 1 }
        temporary_password: { type: string, format: password, minLength: 1 }
        role: { type: string, enum: [SHELTER_ADMIN, STAFF, VOLUNTEER] }
    ShelterArea:
      type: object
      required: [id, organization_id, name, area_type, status]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        area_type: { type: string, enum: [area, cage] }
        parent_id: { type: [string, "null"], format: uuid }
        status: { type: string, enum: [active, inactive] }
    ShelterAreaCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        area_type: { type: string, enum: [area, cage], default: area }
        parent_id: { type: [string, "null"], format: uuid }
    ShelterAreaUpdateRequest:
      type: object
      properties:
        name: { type: string }
        area_type: { type: string, enum: [area, cage] }
        parent_id: { type: [string, "null"], format: uuid }
        status: { type: string, enum: [active, inactive] }
    ShelterAreaListResponse:
      type: object
      required: [items]
      properties:
        items:
          { type: array, items: { $ref: "#/components/schemas/ShelterArea" } }
...